refactor(helper): :recycle: update database helpers
use zod as a checker and to strip out extra fields when parsing. This removes the need for custom checkers and destructure function
