+++ src/helpers/database/papr/audible/PaprAudibleAuthorHelper.ts
@@ -1,12 +1,12 @@
 import AuthorModel, { AuthorDocument } from '#config/models/Author'
-import { isAuthorDocument, isAuthorProfile } from '#config/typing/checkers'
+import { ApiAuthorProfile, ApiAuthorProfileSchema } from '#config/types'
+import { isAuthorDocument } from '#config/typing/checkers'
 import {
 	PaprAuthorDocumentReturn,
 	PaprAuthorReturn,
 	PaprAuthorSearch,
 	PaprDeleteReturn
 } from '#config/typing/papr'
-import { AuthorProfile } from '#config/typing/people'
 import { ParsedQuerystring } from '#config/typing/requests'
 import getErrorMessage from '#helpers/utils/getErrorMessage'
 import SharedHelper from '#helpers/utils/shared'
@@ -21,7 +21,7 @@ import {
 
 export default class PaprAudibleAuthorHelper {
 	asin: string
-	authorData!: AuthorProfile
+	authorData!: ApiAuthorProfile
 	options: ParsedQuerystring
 	sharedHelper = new SharedHelper()
 
@@ -125,10 +125,10 @@ export default class PaprAudibleAuthorHelper {
 			$or: [{ region: { $exists: false } }, { region: this.options.region }]
 		})
 
-		// Remove database fields from data
-		const destructured = this.sharedHelper.destructureDocument(findOneAuthor)
-		// Assign type to author data
-		const data: AuthorProfile | null = isAuthorProfile(destructured) ? destructured : null
+		// Parse data to ensure it's the correct type and remove any extra fields
+		const dataParsed = ApiAuthorProfileSchema.safeParse(findOneAuthor)
+		// Assign data to variable if it's valid, otherwise assign null
+		const data = dataParsed.success ? dataParsed.data : null
 
 		return {
 			data: data,
@@ -139,7 +139,7 @@ export default class PaprAudibleAuthorHelper {
 	/**
 	 * Set authorData in the class object
 	 */
-	setAuthorData(authorData: AuthorProfile) {
+	setAuthorData(authorData: ApiAuthorProfile) {
 		this.authorData = authorData
 	}
